Test correctly
`check-pred` is for testing one value against a predicate.
Currently, it checks that `(< start)` is true (which will always be the
case), and `end` becomes a message, which is probably not
what's intended.

This PR switches to use `check` instead, which can be used to test
a binary relation on two values.